[gaim-migrate @ 11048] i just got an athlon 64, and gcc started bitching.  this should shut it up.

+++ src/protocols/msn/slp.c
@@ -250,7 +250,7 @@ got_sessionreq(MsnSlpCall *slpcall, const char *branch,
 		const char *sha1c;
 		const char *file_name;
 		char *content;
-		gsize len;
+		int len;
 		int type;
 
 		/* Send Ok */
@@ -315,7 +315,7 @@ got_sessionreq(MsnSlpCall *slpcall, const char *branch,
 		GaimAccount *account;
 		GaimXfer *xfer;
 		char *bin;
-		gsize bin_len;
+		int bin_len;
 		guint32 file_size;
 		char *file_name;
